TSK-1989: added tests for jobscheduler
blub

+++ lib/taskana-core/src/main/java/pro/taskana/common/internal/jobs/JobRunner.java
@@ -9,7 +9,6 @@ import org.slf4j.LoggerFactory;
 
 import pro.taskana.common.api.ScheduledJob;
 import pro.taskana.common.api.TaskanaEngine;
-import pro.taskana.common.api.exceptions.SystemException;
 import pro.taskana.common.internal.JobServiceImpl;
 import pro.taskana.common.internal.transaction.TaskanaTransactionProvider;
 
@@ -44,17 +43,20 @@ public class JobRunner {
     TaskanaTransactionProvider.executeInTransactionIfPossible(
         txProvider,
         () -> {
-          taskanaEngine.runAsAdmin(() -> runScheduledJob(scheduledJob));
-          jobService.deleteJob(scheduledJob);
+          Boolean successful = taskanaEngine.runAsAdmin(() -> runScheduledJob(scheduledJob));
+          if (successful) {
+            jobService.deleteJob(scheduledJob);
+          }
         });
   }
 
-  private void runScheduledJob(ScheduledJob scheduledJob) {
+  private boolean runScheduledJob(ScheduledJob scheduledJob) {
     try {
       AbstractTaskanaJob.createFromScheduledJob(taskanaEngine, txProvider, scheduledJob).run();
+      return true;
     } catch (Exception e) {
       LOGGER.error("Error running job: {} ", scheduledJob.getType(), e);
-      throw new SystemException(String.format("Error running job '%s'", scheduledJob.getType()), e);
+      return false;
     }
   }
